FEATURE(Menu): 원래의 menu를 request로 수정하고 Menu도메인 객체 생성
